Make script (mostly) posix-compliant
Remove all the bashisms so the script will run in most
posix shells. The only non-posix thing left is local variables,
which are supported in almost all shells.

Change exit code -1 to 2, since -1 is invalid, and is converted
to 255 (which means invalid exit code).

Fix comparison of version w/ special with version w/o special,
which was failing the test.

Also make test script return exit status 1 if a test fails.